+++ src/braket/analog_hamiltonian_simulator/rydberg/validators/times_series.py
@@ -14,7 +14,7 @@
 import warnings
 
 from braket.ir.ahs.time_series import TimeSeries
-from pydantic.class_validators import root_validator
+from pydantic.v1.class_validators import root_validator
 
 from braket.analog_hamiltonian_simulator.rydberg.validators.capabilities_constants import (
     CapabilitiesConstants,
@@ -63,7 +63,7 @@ class TimeSeriesValidator(TimeSeries):
         for i in range(len(times) - 1):
             if times[i] >= times[i + 1]:
                 raise ValueError(
-                    f"Time point {i} ({times[i]}) and time point {i+1} ({times[i+1]}) "
+                    f"Time point {i} ({times[i]}) and time point {i + 1} ({times[i + 1]}) "
                     "must be monotonically increasing."
                 )
         return values