+++ test/unit_tests/braket/default_simulator/openqasm/test_interpreter.py
@@ -22,7 +22,7 @@ import sympy
 from sympy import Symbol
 
 from braket.default_simulator import StateVectorSimulation
-from braket.default_simulator.gate_operations import CX, Hadamard, PauliX
+from braket.default_simulator.gate_operations import CX, GPhase, Hadamard, PauliX
 from braket.default_simulator.gate_operations import PauliY as Y
 from braket.default_simulator.gate_operations import RotX, U, Unitary
 from braket.default_simulator.noise_operations import (
@@ -115,6 +115,8 @@ def test_int_declaration():
     int[8] uninitialized;
     int[8] pos = 10;
     int[5] neg = -4;
+    int[8] int_min = -128;
+    int[8] int_max = 127;
     int[3] pos_overflow = 5;
     int[3] neg_overflow = -6;
     int no_size = 1e9;
@@ -135,8 +137,10 @@ def test_int_declaration():
     assert context.get_value("uninitialized") is None
     assert context.get_value("pos") == IntegerLiteral(10)
     assert context.get_value("neg") == IntegerLiteral(-4)
-    assert context.get_value("pos_overflow") == IntegerLiteral(1)
-    assert context.get_value("neg_overflow") == IntegerLiteral(-2)
+    assert context.get_value("int_min") == IntegerLiteral(-128)
+    assert context.get_value("int_max") == IntegerLiteral(127)
+    assert context.get_value("pos_overflow") == IntegerLiteral(-3)
+    assert context.get_value("neg_overflow") == IntegerLiteral(2)
     assert context.get_value("no_size") == IntegerLiteral(1_000_000_000)
 
     warnings = {(warn.category, warn.message.args[0]) for warn in warn_info}
@@ -173,6 +177,34 @@ def test_uint_declaration():
     assert context.get_value("pos_overflow") == IntegerLiteral(0)
     assert context.get_value("neg_overflow") == IntegerLiteral(7)
     assert context.get_value("no_size") == IntegerLiteral(1_000_000_000)
+
+
+def test_signed_int_cast():
+    qasm = """
+    uint[8] x0 = 255;
+    int[8] x1 = x0;
+    uint[8] x2 = x1;
+
+    uint[8] y0 = 128;
+    int[8] y1 = y0;
+    uint[8] y2 = y1;
+
+    int[3] z0 = "100";
+    int[3] z1 = "111";
+    """
+
+    context = Interpreter().run(qasm)
+
+    assert context.get_value("x0") == IntegerLiteral(255)
+    assert context.get_value("x1") == IntegerLiteral(-1)
+    assert context.get_value("x2") == IntegerLiteral(255)
+
+    assert context.get_value("y0") == IntegerLiteral(128)
+    assert context.get_value("y1") == IntegerLiteral(-128)
+    assert context.get_value("y2") == IntegerLiteral(128)
+
+    assert context.get_value("z0") == IntegerLiteral(-4)
+    assert context.get_value("z1") == IntegerLiteral(-1)
 
 
 def test_float_declaration():
@@ -685,7 +717,7 @@ def test_update_bits_int():
     """
     context = Interpreter().run(qasm)
     assert context.get_value("x") == IntegerLiteral(3)
-    assert context.get_value("y") == IntegerLiteral(-2)
+    assert context.get_value("y") == IntegerLiteral(-6)
     assert context.get_value("z") == IntegerLiteral(10)
 
 
@@ -888,8 +920,21 @@ def test_gate_inv():
     inv @ t q;
     """
     circuit = Interpreter().build_circuit(qasm)
-    collapsed = np.linalg.multi_dot([instruction.matrix for instruction in circuit.instructions])
-    assert np.allclose(collapsed, np.eye(2**circuit.num_qubits))
+    coeff = np.linalg.multi_dot(
+        [
+            instruction.matrix
+            for instruction in circuit.instructions
+            if isinstance(instruction, GPhase)
+        ]
+    )[0][0]
+    collapsed = np.linalg.multi_dot(
+        [
+            instruction.matrix
+            for instruction in circuit.instructions
+            if not isinstance(instruction, GPhase)
+        ]
+    )
+    assert np.allclose(coeff * collapsed, np.eye(2**circuit.num_qubits))
 
 
 def test_gate_ctrl():
@@ -1015,7 +1060,7 @@ def test_pow():
         pow(two) @ cx c, a;
     }
     gate cxx_2 c, a {
-        pow(1/2) @ pow(4) @ cx c, a;
+        pow(1./2.) @ pow(4) @ cx c, a;
     }
     gate cxxx c, a {
         pow(1) @ pow(two) @ cx c, a;
@@ -1027,18 +1072,19 @@ def test_pow():
     qubit q4;
     qubit q5;
 
-    pow(1/2) @ x q1;       // half flip
-    pow(1/2) @ x q1;       // half flip
-    cx q1, q2;   // flip
-    cxx_1 q1, q3;    // don't flip
-    cxx_2 q1, q4;    // don't flip
-    cnot q1, q5;    // flip
-    x q3;       // flip
-    x q4;       // flip
-
-    s q1;   // sqrt z
-    s q1;   // again
-    inv @ z q1; // inv z
+    pow(1./2) @ x q1;   // half flip
+    pow(1/2.) @ x q1;  // half flip
+    cx q1, q2;          // flip
+    cxx_1 q1, q3;       // don't flip
+    cxx_2 q1, q4;       // don't flip
+    cnot q1, q5;        // flip
+    x q3;               // flip
+    x q4;               // flip
+    pow(1/2) @ x q5;    // don't flip
+    
+    s q1;               // sqrt z
+    s q1;               // again
+    inv @ z q1;         // inv z
     """
     circuit = Interpreter().build_circuit(qasm)
     simulation = StateVectorSimulation(5, 1, 1)
@@ -2055,3 +2101,220 @@ def test_basis_rotation_hermitian():
         Hadamard([2]),
         *hermitian.diagonalizing_gates(),
     ]
+
+
+@pytest.mark.parametrize(
+    "qasm, expected",
+    [
+        (
+            "\n".join(
+                [
+                    "bit[1] b;",
+                    "qubit[2] q;",
+                    "h q[0];",
+                    "h q[1];",
+                    "b[0] = measure q[0];",
+                ]
+            ),
+            ([0], [0]),
+        ),
+        (
+            "\n".join(
+                [
+                    "bit[3] b;",
+                    "qubit[3] q;",
+                    "b = measure q;",
+                ]
+            ),
+            ([0, 1, 2], [0, 1, 2]),
+        ),
+        (
+            "\n".join(
+                [
+                    "bit[2] b;",
+                    "qubit[2] q;",
+                    "h q[0];",
+                    "h q[1];",
+                    "b[0:1] = measure q[0:1];",
+                ]
+            ),
+            ([0, 1], [0, 1]),
+        ),
+        (
+            "\n".join(
+                [
+                    "bit[3] b;",
+                    "qubit[3] q;",
+                    "h q[0];",
+                    "cnot q[0], q[1];",
+                    "cnot q[1], q[2];",
+                    "b[0] = measure q[0];",
+                    "b[2] = measure q[1];",
+                    "b[1] = measure q[2];",
+                ]
+            ),
+            ([0, 1, 2], [0, 2, 1]),
+        ),
+        (
+            "\n".join(
+                [
+                    "bit[1] b;",
+                    "qubit[3] q;",
+                    "h q[0];",
+                    "h q[1];",
+                    "cnot q[1], q[2];",
+                    "b[{2, 1}] = measure q[{0, 2}];",
+                ]
+            ),
+            ([0, 2], [2, 1]),
+        ),
+        (
+            "\n".join(
+                [
+                    "bit[1] b;",
+                    "h $0;",
+                    "cnot $0, $1;",
+                    "b[0] = measure $0;",
+                ]
+            ),
+            ([0], [0]),
+        ),
+        (
+            "\n".join(
+                [
+                    "qubit[5] q;",
+                    "for int i in [0:2] {",
+                    "   measure q[i];",
+                    "}",
+                ]
+            ),
+            ([0, 1, 2], [0, 1, 2]),
+        ),
+        (
+            "\n".join(
+                [
+                    "bit[1] b;",
+                    "qubit[3] q;",
+                    "h q[0];",
+                    "h q[1];",
+                    "cnot q[1], q[2];",
+                    "measure q[1];",
+                    "measure q[0];",
+                ]
+            ),
+            ([1, 0], [0, 1]),
+        ),
+        (
+            "\n".join(
+                [
+                    "bit[1] b;",
+                    "qubit[2] q;",
+                    "b[0] = measure q[1:5];",
+                ]
+            ),
+            ([1], [0]),
+        ),
+    ],
+)
+def test_measurement(qasm, expected):
+    circuit = Interpreter().build_circuit(qasm)
+    assert circuit.measured_qubits == expected[0]
+    assert circuit.target_classical_indices == expected[1]
+
+
+@pytest.mark.parametrize(
+    "qasm, expected",
+    [
+        (
+            "\n".join(
+                [
+                    "bit[3] b;",
+                    "qubit[2] q;",
+                    "h q[0];",
+                    "cnot q[0], q[1];",
+                    "b[2] = measure q[1];",
+                    "b[0] = measure q[0];",
+                    "b[1] = measure q[0];",
+                ]
+            ),
+            "Qubit 0 is already measured or captured.",
+        ),
+        (
+            "\n".join(
+                [
+                    "bit[1] b;",
+                    "qubit[1] q;",
+                    "h q[0];",
+                    "b[0] = measure q[0];",
+                    "measure q;",
+                ]
+            ),
+            "Qubit 0 is already measured or captured.",
+        ),
+    ],
+)
+def test_measurement_exceptions(qasm, expected):
+    with pytest.raises(ValueError, match=expected):
+        Interpreter().build_circuit(qasm)
+
+
+def test_measure_invalid_qubit():
+    qasm = """
+    bit[1] b;
+    qubit[1] q;
+    h q[0];
+    measure x;
+    """
+    expected = "Undefined key: x"
+    with pytest.raises(KeyError, match=expected):
+        Interpreter().build_circuit(qasm)
+
+
+@pytest.mark.parametrize(
+    "qasm, expected",
+    [
+        (
+            "\n".join(
+                [
+                    "bit[1] b;",
+                    "qubit[1] q;",
+                    "b[0] = measure q[5];",
+                ]
+            ),
+            "qubit register index `5` out of range for qubit register of length 1 `q`.",
+        ),
+        (
+            "\n".join(
+                [
+                    "bit[1] b;",
+                    "qubit[2] q;",
+                    "b[0] = measure q[{1, 5}];",
+                ]
+            ),
+            "qubit register index `5` out of range for qubit register of length 2 `q`.",
+        ),
+    ],
+)
+def test_measure_qubit_out_of_range(qasm, expected):
+    with pytest.raises(IndexError, match=expected):
+        Interpreter().build_circuit(qasm)
+
+
+@pytest.mark.parametrize(
+    "qasm,error_message",
+    [
+        (
+            "\n".join(["OPENQASM 3.0;" "bit[2] b;", "qubit[1] q;", "b[{0, 1}] = measure q[0];"]),
+            re.escape(
+                "Number of qubits (1) does not match number of provided classical targets (2)"
+            ),
+        ),
+        (
+            "\n".join(["OPENQASM 3.0;" "bit[2] b;", "qubit[2] q;", "b[0][2] = measure q[1];"]),
+            re.escape("Multi-Dimensional indexing not supported for classical registers."),
+        ),
+    ],
+)
+def test_invalid_measurement_with_classical_indices(qasm, error_message):
+    with pytest.raises(ValueError, match=error_message):
+        Interpreter().build_circuit(qasm)